requirements_test.txt

-r requirements.txt
prettytable==3.12.0
async-timeout
asynctest~=0.13
colorlog~=6.8
flake8~=7.0
flake8-docstrings~=1.7
mypy~=1.10
pylint~=3.2
pylint-strict-informational==0.1
pytest>=7.2
pytest-cov>=3.0
pytest-homeassistant-custom-component>=0.13
tzdata
ruff>=0.4
josepy<2.0
